+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/core/KMaP.kt
@@ -10,9 +10,9 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import com.rafambn.kmap.tiles.TileCanvas
-import com.rafambn.kmap.lazyMarker.KMaPScope
-import com.rafambn.kmap.lazyMarker.rememberComponentProviderLambda
-import com.rafambn.kmap.lazyMarker.rememberComponentMeasurePolicy
+import com.rafambn.kmap.lazy.KMaPScope
+import com.rafambn.kmap.lazy.rememberComponentProviderLambda
+import com.rafambn.kmap.lazy.rememberComponentMeasurePolicy
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
