+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/lazy/KMaPScope.kt
@@ -1,4 +1,4 @@
-package com.rafambn.kmap.lazyMarker
+package com.rafambn.kmap.lazy
 
 import androidx.compose.foundation.lazy.LazyScopeMarker
 import androidx.compose.runtime.Composable
