+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/lazy/MeasuredComponent.kt
@@ -1,4 +1,4 @@
-package com.rafambn.kmap.lazyMarker
+package com.rafambn.kmap.lazy
 
 import androidx.compose.ui.layout.Placeable
 import com.rafambn.kmap.components.MarkerParameters
